redundant coverartlabel file deleted.
