+++ CMakeLists.txt
@@ -688,7 +688,6 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/library/coverartdelegate.cpp
   src/library/coverartutils.cpp
   src/library/dlgcoverartfetchedfullsize.cpp
-  src/library/coverartlabel.cpp
   src/library/dao/analysisdao.cpp
   src/library/dao/autodjcratesdao.cpp
   src/library/dao/cuedao.cpp
